feat: emotion theme
