+++ src/pages/_app.tsx
@@ -1,11 +1,13 @@
 import type { AppContext, AppProps } from 'next/app';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
+import { ThemeProvider } from '@emotion/react';
 import { domMax, LazyMotion } from 'framer-motion';
 
 import { BASE_URL } from '~/constant/common';
 import { useRecordPageView } from '~/hooks/useRecordPageView';
 import GlobalStyle from '~/styles/globalStyle';
+import { theme } from '~/styles/theme';
 
 interface InitialProps {
   userAgent: string;
@@ -18,15 +20,16 @@ export default function App({ Component, pageProps }: AppProps & InitialProps) {
   useRecordPageView();
 
   return (
-    <LazyMotion features={domMax}>
-      <Head>
-        <link rel="canonical" href={currentUrl} />
-        <meta property="og:url" content={currentUrl} />
-      </Head>
-
-      <GlobalStyle />
-      <Component {...pageProps} />
-    </LazyMotion>
+    <ThemeProvider theme={theme}>
+      <LazyMotion features={domMax}>
+        <Head>
+          <link rel="canonical" href={currentUrl} />
+          <meta property="og:url" content={currentUrl} />
+        </Head>
+        <GlobalStyle />
+        <Component {...pageProps} />
+      </LazyMotion>
+    </ThemeProvider>
   );
 }
 App.getInitialProps = async ({ ctx }: AppContext) => {
