+++ src/styles/colors.ts
@@ -11,5 +11,5 @@ export const colors = {
   blue300: '#0973EE',
   yellow: '#E3FF3A',
   yellow400: '#C6E413',
-  yellow300: '##A1BB00',
+  yellow300: '#A1BB00',
 } as const;
