+++ src/styles/theme.ts
@@ -1,0 +1,7 @@
+import { colors } from './colors';
+// import { typos } from './typo';
+
+export const theme = {
+  colors,
+  // typos,
+} as const;
